+++ service/src/main/java/org/whispersystems/signalservice/internal/push/PushServiceSocket.java
@@ -619,7 +619,7 @@ public class PushServiceSocket implements Closeable {
     return JsonUtil.fromJson(responseText, DeviceInfoList.class).getDevices();
   }
 
-  public void removeDevice(long deviceId) throws IOException {
+  public void removeDevice(int deviceId) throws IOException {
     makeServiceRequest(String.format(DEVICE_PATH, String.valueOf(deviceId)), "DELETE", null);
   }
